Refactor CSM Observability Test logic into Mixin class
